+++ src/main/java/tfw/immutable/ila/floatila/FloatIlaDecimate.java
@@ -44,7 +44,7 @@ public final class FloatIlaDecimate {
             final FloatIlaIterator fi = new FloatIlaIterator(segment, bufferSize);
 
             for (int ii = offset; length > 0; ii += stride, --length) {
-                array[ii] = (float) fi.next();
+                array[ii] = fi.next();
                 fi.skip(factor - 1);
             }
         }
